chore: move endpoitn to new OnboardingController
