+++ server/src/main/resources/togglz.features-file.properties
@@ -18,3 +18,5 @@ REVISION_ASSIGNMENT=true
 
 # Hide the OIDC login button but keep the OIDC login flow enabled
 SHOW_OIDC_LOGIN=true
+
+ONBOARDING=true
